### 2023-04-12 #5360 @editorialbot
[REVIEW]: r2ogs6: An R wrapper of the OpenGeoSys 6 Multiphysics Simulator
**Submitting author:** @joboog (<a href="http://orcid.org/0000-0003-1094-9255">Johannes Boog</a>)
**Repository:** https://github.com/joboog/r2ogs6
**Branch with paper.md** (empty if default branch): 40-joss-submission
**Version:** v0.4.652
