### 2023-06-13 #5360 @joboog
@waternumbers Thanks for your review. I have also a push mirror at https://github.com/joboog/r2ogs6 where you could open an issue if that would be ok with the JOSS guidelines (@jsta ).
