### 2023-06-30 af553f2
[base] add python packages as deps
