+++ DESCRIPTION
@@ -26,6 +26,11 @@ Description: An R-API to the OpenGeoSys-v6 multi-physics simulator.
 License: GPL-3
 Encoding: UTF-8
 LazyData: true
+Config/reticulate:
+  list(
+    packages = list(
+      list(package = "numpy", version = "1.24.2", pip = TRUE),
+      list(package = "vtk", version = "9.2.6", pip = TRUE)))
 Suggests: 
     covr,
     testthat,
