### 2023-06-30 73e9d92
[base] move biocmanager and rhdf5 to imports
