+++ DESCRIPTION
@@ -37,9 +37,8 @@ Suggests:
     knitr,
     devtools,
     ggplot2,
-    BiocManager,
-    rhdf5,
     rmarkdown
+biocViews:
 Imports: 
     xml2,
     tibble,
@@ -52,7 +51,9 @@ Imports:
     doParallel,
     reticulate,
     config,
-    foreach
+    foreach,
+    BiocManager,
+    rhdf5
 RoxygenNote: 7.2.3
 VignetteBuilder: knitr
 Depends: 
