### 2023-06-30 66ac8ed
[base] do not check for biocmanager and rhdf5 deps at startup
