### 2023-07-04 a6c9a0b
remove python deps from DESCRIPTION
