### 2023-07-08 9982e3f
[JOSS] add references to 1st paragraph of summary
