+++ JOSS/paper.bib
@@ -1,3 +1,32 @@
+@article{Felfelani2017,
+  title={Natural and human-induced terrestrial water storage change: A global analysis using hydrological models and GRACE},
+  author={Felfelani, Farshid and Wada, Yoshihide and Longuevergne, Laurent and Pokhrel, Yadu N},
+  journal={Journal of Hydrology},
+  volume={553},
+  pages={105--118},
+  year={2017},
+  publisher={Elsevier}
+}
+
+@article{Wu2020,
+  title={Divergent effects of climate change on future groundwater availability in key mid-latitude aquifers},
+  author={Wu, Wen-Ying and Lo, Min-Hui and Wada, Yoshihide and Famiglietti, James S and Reager, John T and Yeh, Pat J-F and Ducharne, Agn{\`e}s and Yang, Zong-Liang},
+  journal={Nature communications},
+  volume={11},
+  number={1},
+  pages={3710},
+  year={2020},
+  publisher={Nature Publishing Group UK London}
+}
+
+@article{Li2023,
+  title={Big Data in Earth system science and progress towards a digital twin},
+  author={Li, Xin and Feng, Min and Ran, Youhua and Su, Yang and Liu, Feng and Huang, Chunlin and Shen, Huanfeng and Xiao, Qing and Su, Jianbin and Yuan, Shiwei and others},
+  journal={Nature Reviews Earth \& Environment},
+  pages={1--14},
+  year={2023},
+  publisher={Nature Publishing Group UK London}
+}
 %------------------ OpenGeoSys
 
 @article{Steefel2015,
