+++ JOSS/paper.bib
@@ -67,12 +67,16 @@ address = {Cham}
   isbn={978-3-642-27176-2}
 }
 
-@Manual{Buchwald2021,
-  author = {Buchwald, J{\"o}rg and Fischer, Tom and Naumov, Dmitry Yu and Kolditz, Olaf},
-  title = {ogs6py: Python-API for the OpenGeoSys Software},
-  year = {2021},
-  url = {https://github.com/joergbuchwald/ogs6py},
-  doi={10.5281/zenodo.5387765}
+@article{Buchwald2021,
+  doi = {10.21105/joss.03673},
+  url = {https://doi.org/10.21105/joss.03673},
+  year = {2021}, publisher = {The Open Journal},
+  volume = {6},
+  number = {67},
+  pages = {3673},
+  author = {Jörg Buchwald and Olaf Kolditz and Thomas Nagel},
+  title = {ogs6py and VTUinterface: streamlining OpenGeoSys workflows in Python},
+  journal = {Journal of Open Source Software}
 }
 
 %------------------ r
@@ -219,8 +223,44 @@ year = {2021}
   journal = {Journal of Open Source Software}
 }
 
-@article{Schad2021,
-author = {Schad, Philipp and Boog, Johannes and Kalbacher, Thomas},
-title = {r2ogs5: Controlling and calibrating OpenGeoSys 5 simulation within R},
-journal = {Submitted to Groundwater},
-year = {2021}}
+@article{Schad2023,
+  title={r2ogs5: Calibration of Numerical Groundwater Flow Models with Bayesian Optimization in R},
+  author={Schad, Philipp and Boog, Johannes and Kalbacher, Thomas},
+  journal={Groundwater},
+  volume={61},
+  number={1},
+  pages={119--130},
+  year={2023},
+  publisher={Wiley Online Library}
+}
+
+@article{Langevin2021,
+  title={MODFLOW 6 modular hydrologic model version 6.2. 2},
+  author={Langevin, CD and Hughes, JD and Banta, ER and Provost, AM and Niswonger, RG and Panday, Sorab},
+  journal={US Geological Survey Software Release},
+  volume={10},
+  pages={F76Q1VQV},
+  year={2021}
+}
+
+@article{Pruess2004,
+  title={The TOUGH codes—A family of simulation tools for multiphase flow and transport processes in permeable media},
+  author={Pruess, Karsten},
+  journal={Vadose zone journal},
+  volume={3},
+  number={3},
+  pages={738--746},
+  year={2004},
+  publisher={Soil Science Society of America}
+}
+
+@article{Charlton2011,
+  title={Modules based on the geochemical model PHREEQC for use in scripting and programming languages},
+  author={Charlton, Scott R and Parkhurst, David L},
+  journal={Computers \& Geosciences},
+  volume={37},
+  number={10},
+  pages={1653--1663},
+  year={2011},
+  publisher={Elsevier}
+}
