+++ JOSS/paper.md
@@ -63,11 +63,17 @@ But while `OpenGeoSys` is a powerful FEM code, setting up, running and evaluatin
 
 Here is where languages such as `R` and `Python` can prove useful.
 Via an interface that adds a layer on top of `OpenGeoSys 6`, the user can access preprocessing tools, the solver itself and postprocessing tools alike, thus increasing usability and accessibility.
-The developement and application of user interfaces from geoscientific simulators to high--level programming languages has been gaining increasing attention in recent years; examples are `FloPy` [@Bakker2016], `ogs5py` [@Mueller2021], `RedModRPhree`[@DeLucia2021],  `r2ogs5` [@Schad2021] and `toughio` [@Luu2020].
-For `OpenGeoSys 6`, a `Python` interface is currently under development [@Buchwald2021]. 
-Nonetheless, we consider an `R` interface to be just as important, as `R` is a well known language in the environmental and geosciences.
+The development and application of high--level programming and/or scripting languages interfaces to geoscientific simulators   has been gaining increasing attention in recent years. Examples are a) `FloPy` [@Bakker2016]: a `Python` interface to the groundwater simulator `Modflow` [@Langevin2021],  `RedModRPhree` [@DeLucia2021]: `R` utility functions to the geo-chemical solver `Phreeqc` [@Charlton2011], the `toughio` [@Luu2020] Python interface to the multi-phase flow simulator `TOUGH`  [@Pruess2004], the Python interfaces to `OpenGeoSys 5`: `ogs5py` [@Mueller2021] and `OpenGeoSys 6`: `ogs6py` [@Buchwald2021] as well as the `R` interface to `OpenGeoSys 5`: `r2ogs5` [@Schad2023]. 
+`FloPy`,  `ogs5py`, `ogs6py` and `r2ogs5` provide object-oriented frameworks to set-up, call the executables and import the output of the associated simulators.
+The idea of these packages is to provide `Python` and `R` functions to pre- and process simulation data but to just run simulations via a call to  the the operating system to execute the simulators with the prepared input.
+`RedModRPhree` and `toughio` rather focus on additional pre- and post processing functions to simplify the set-up of simulations with `R` and `Python`.
+These packages still rely on external execution of the associated simulation program.
+
+Altough there already exist the `OpenGeoSys 6` `Python` interface `ogs6py`, we consider an `R` interface to be just as important, as `R` is a well known language in the environmental and geosciences.
 Furthermore, since `R` is a popular language in the field of data science with many powerful packages for data analysis and visualisation, e. g. `dplyr` [@r-dplyr] and `ggplot2` [@r-ggplot2], it's a natural choice for processing data generated by simulation tools such as `OpenGeoSys 6`.
-Especially, `r2ogs6` can facilitate the calibration of `OpenGeoSys 6` models due to the implemented functions to design ensemble runs as well the available `R` functions and packages for modeling such as  `lhs` [@lhs], `mlrBO` [@mlrMBO].
+`r2ogs6` follows a similar design approach compared to `FloPy`,  `ogs5py`, `ogs6py` and `r2ogs5`: it provides an object-oriented approach with `R` functions to set-up an `OpenGeoSys 6` simulation, call the respective executable and import the output of finished simulations (see section *Package Structure* for more details).
+
+Moreover, `r2ogs6` can facilitate the calibration of `OpenGeoSys 6` models due to the implemented functions to design ensemble runs as well the available `R` functions and packages for modeling such as  `lhs` [@lhs], `mlrBO` [@mlrMBO].
 For `R` users who do not have a lot of (or any) experience with, yet an interest in environmental and geoscientific sub--surface simulations, `r2ogs6` provides a good starting point.
[…]
