### 2023-07-08 cd87d57
[JOSS] improve clarity of the paper
