+++ JOSS/paper.md
@@ -33,8 +33,9 @@ bibliography: paper.bib
 Understanding the impacts of climate change and hydrologic extreme events on our sub--surface earth system is even in temperate zones of utmost importance [@Felfelani2017; @Wu2020].
 Key tools to develop such understanding are physics--based simulation models that describe the manifold interactions of involved natural phenomena across time and space [@Li2023; @Steefel2015].
 
-Our `R` package `r2ogs6` provides a file--based `R` interface to the multi--physics simulation code `OpenGeoSys 6` [@Kolditz2012; @Bilke2019] and, therefore, enables `R` users to perform and analyze simulation models of the sub--surface earth system in `R`.
-`r2ogs6` allows to access the capabilities of `OpenGeoSys 6` to simulate thermo-hydro-mechanical-chemical and biological (THMC/B) processes in porous and fractured media within `R`.
+Our package `r2ogs6` provides a file--based interface of the multi--physics simulation code `OpenGeoSys 6` [@Kolditz2012; @Bilke2019] to the free software environment for statistical computing `R`.
+`r2ogs6` enables `R` users to perform and analyze simulation models of the sub--surface earth system in `R`.
+It allows to access the capabilities of `OpenGeoSys 6` to simulate thermo-hydro-mechanical-chemical and biological (THMC/B) processes in porous and fractured media within `R`.
 In this way, `r2ogs6` enables `R` users to model sub--surface phenomena and technologies such as groundwater flow, reactive transport, geothermal energy usage and/or nuclear waste repositories as well as to analyze and further process simulation output.
 `r2ogs6` enables users to prepare and manipulate `OpenGeoSys 6` simulation models, run the simulations and retrieve corresponding output, all within an `R` session or simple `R` scripts.
 Therefore, `R` classes and functions were designed to communicate with the respective `OpenGeoSys 6` input and output files as well as executables.
@@ -58,7 +59,9 @@ For reasons of performance, (multi) physics simulators are mostly implemented in
 
 One of these tools is `OpenGeoSys` (OGS) (https://www.opengeosys.org/), a scientific open source project for the development of numerical methods to simulate thermo-hydro-mechanical-chemical and biological (THMC/B) processes in porous and fractured media [@Kolditz2012; @Bilke2019].
 OGS has applications ranging from small-scale geotechnical investigations, to reservoir studies and even groundwater management of entire landscapes.  
-Since the model setups get more and more complex and larger, faster and more efficient setup and parametrization procedures are needed to create meaningful ensembles that allow to identify the right action needs and to derive further decision support for the above described problems.
+To identify the right action needs and to derive further decision support for the above described problems it often requires to explore several problem scenarios through meaningful model ensembles.
+This translates to more and more complex and larger model setups.
+Thus, faster and more efficient model setup and parametrization procedures are needed.
 But while `OpenGeoSys` is a powerful FEM code, setting up, running and evaluating multiple simulations can prove complicated.
 
 Here is where languages such as `R` and `Python` can prove useful.
@@ -73,7 +76,9 @@ Altough there already exist the `OpenGeoSys 6` `Python` interface `ogs6py`, we c
 Furthermore, since `R` is a popular language in the field of data science with many powerful packages for data analysis and visualisation, e. g. `dplyr` [@r-dplyr] and `ggplot2` [@r-ggplot2], it's a natural choice for processing data generated by simulation tools such as `OpenGeoSys 6`.
[…]
