### 2023-07-09 3a5a5c2
[JOSS] updte reference of lhs to point to cran
