+++ JOSS/paper.bib
@@ -70,11 +70,12 @@ address = {Cham}
 @article{Buchwald2021,
   doi = {10.21105/joss.03673},
   url = {https://doi.org/10.21105/joss.03673},
-  year = {2021}, publisher = {The Open Journal},
+  year = {2021},
+  publisher = {The Open Journal},
   volume = {6},
   number = {67},
   pages = {3673},
-  author = {Jörg Buchwald and Olaf Kolditz and Thomas Nagel},
+  author = {J{\"o}rg Buchwald and Olaf Kolditz and Thomas Nagel},
   title = {ogs6py and VTUinterface: streamlining OpenGeoSys workflows in Python},
   journal = {Journal of Open Source Software}
 }
@@ -113,9 +114,9 @@ address = {Cham}
   title = {lhs: Latin Hypercube Samples},
   author = {Rob Carnell},
   year = {2021},
-  note = {R package version 1.1.4},
-  url = {https://bertcarnell.github.io/lhs/index.html},
-  urldate = {2021-10-04}
+  note = {R package version 1.1.6},
+  url = {https://cran.r-project.org/package=lhs},
+  urldate = {2023-07-09}
 }
 
 @article{GPfit,
