### 2023-07-09 8d2201a
[JOSS] further improve clarity of the paper
