+++ JOSS/paper.md
@@ -12,28 +12,29 @@ authors:
   - name: Johannes Boog^[co-first author]
     orcid: 0000-0003-0872-7098
     affiliation: 2
-  - name: Philipp Schad^[co-first author]  
-    orcid: 0000-0003-3332-5867  
-    affiliation: 2  
+  - name: Philipp Schad^[co-first author]
+    orcid: 0000-0003-3332-5867
+    affiliation: 2
   - name: Thomas Kalbacher^[co-first author]
     orcid: 0000-0002-7866-5702
     affiliation: 2
 affiliations:
- - name: Leipzig University of Applied Sciences, Karl-Liebknecht-Strasse 132, 04277 Leipzig, Germany
-   index: 1
- - name: Helmholtz Centre for Environmental Research, Department Environmental Informatics, Permoser Str. 15, 04318 Leipzig, Germany
-   index: 2
+  - name: Leipzig University of Applied Sciences, Karl-Liebknecht-Strasse 132, 04277
+      Leipzig, Germany
+    index: 1
+  - name: Helmholtz Centre for Environmental Research, Department Environmental Informatics,
+      Permoser Str. 15, 04318 Leipzig, Germany
+    index: 2
 date: 19 December 2022
 bibliography: paper.bib
-
 ---
 
 # Summary
 
-Understanding the impacts of climate change and hydrologic extreme events on our sub--surface earth system is even in temperate zones of utmost importance [@Felfelani2017; @Wu2020].
+Understanding the impacts of climate change and hydrologic extreme events on our sub--surface earth system is even in temperate zones of utmost importance [@Felfelani2017; @Field2012; @Wu2020].
 Key tools to develop such understanding are physics--based simulation models that describe the manifold interactions of involved natural phenomena across time and space [@Li2023; @Steefel2015].
 
-Our package `r2ogs6` provides a file--based interface of the multi--physics simulation code `OpenGeoSys 6` [@Kolditz2012; @Bilke2019] to the free software environment for statistical computing `R`.
+Our package `r2ogs6` provides a file--based interface of the multi--physics simulation code `OpenGeoSys 6` [@Kolditz2012; @Bilke2019] to the [`R`](www.r-project.org) programing and statistical computing environment.
 `r2ogs6` enables `R` users to perform and analyze simulation models of the sub--surface earth system in `R`.
 It allows to access the capabilities of `OpenGeoSys 6` to simulate thermo-hydro-mechanical-chemical and biological (THMC/B) processes in porous and fractured media within `R`.
 In this way, `r2ogs6` enables `R` users to model sub--surface phenomena and technologies such as groundwater flow, reactive transport, geothermal energy usage and/or nuclear waste repositories as well as to analyze and further process simulation output.
@@ -49,15 +50,15 @@ Moreover, `r2ogs6` was intended to include `OpenGeoSys 6` into `R` based scienti
 
 # Statement of need
 
-Major challenges humanity has to face in the coming decades are climate change and hydrologic extremes.
-Understanding the impacts of climate change and hydrologic extreme events on our sub--surface earth system is even in temperate zones of utmost importance for ensuring adequate domestic and drinking water supplies, together with functioning lake and river systems with healthy aquatic ecosystems and ecosystem services. 
+Major challenges humanity has to face in the coming decades are climate change and hydrologic extremes [@Field2012].
+Understanding the impacts of climate change and hydrologic extreme events on our sub--surface earth system is even in temperate zones of utmost importance for ensuring adequate domestic and drinking water supplies, together with functioning lake and river systems with healthy aquatic ecosystems and ecosystem services [@Felfelani2017; @Wu2020]. 
 Of course, the needs of the population and the needs of nature are often in conflict, which increases the necessity to study the complex interaction of both within different scenarios.
[…]
