+++ JOSS/paper.bib
@@ -1,3 +1,10 @@
+@book{Field2012,
+place={Cambridge},
+title={Managing the Risks of Extreme Events and Disasters to Advance Climate Change Adaptation: Special Report of the Intergovernmental Panel on Climate Change},
+DOI={10.1017/CBO9781139177245},
+publisher={Cambridge University Press},
+year={2012}}
+
 @article{Felfelani2017,
   title={Natural and human-induced terrestrial water storage change: A global analysis using hydrological models and GRACE},
   author={Felfelani, Farshid and Wada, Yoshihide and Longuevergne, Laurent and Pokhrel, Yadu N},
