### 2023-07-09 67600fa
[JOSS] add examples for practical use and script generation from benchmarks
