### 2023-07-09 4024726
[JOSS] chane book references to article format
