+++ JOSS/paper.bib
@@ -1,5 +1,5 @@
-@book{Field2012,
-place={Cambridge},
+@article{Field2012,
+author={Field, C. and Barros, V. and Stocker, T. and Dahe, Q.},
 title={Managing the Risks of Extreme Events and Disasters to Advance Climate Change Adaptation: Special Report of the Intergovernmental Panel on Climate Change},
 DOI={10.1017/CBO9781139177245},
 publisher={Cambridge University Press},
