### 2023-07-09 ad23da0
[JOSS] adjust format for JOSS print
