+++ JOSS/paper.md
@@ -53,8 +53,8 @@ Moreover, `r2ogs6` was intended to include `OpenGeoSys 6` into `R` based scienti
 Major challenges humanity has to face in the coming decades are climate change and hydrologic extremes [@Field2012].
 Understanding the impacts of climate change and hydrologic extreme events on our sub--surface earth system is even in temperate zones of utmost importance for ensuring adequate domestic and drinking water supplies, together with functioning lake and river systems with healthy aquatic ecosystems and ecosystem services [@Felfelani2017; @Wu2020]. 
 Of course, the needs of the population and the needs of nature are often in conflict, which increases the necessity to study the complex interaction of both within different scenarios.
-The core of such studies is most often the system and scenario analysis of individual or coupled earth systems compartments through physics simulations . 
-In physics simulation models multiple coupled natural processes are implemented, which  are usually described with partial differential equations.
+The core of such studies is most often the system and scenario analysis of individual or coupled earth systems compartments through physics simulations. 
+In physics simulation models, multiple coupled natural processes are implemented, which  are usually described with partial differential equations.
 Solving these equations requires appropriate numerical methods such as the finite element method (FEM).
 For reasons of performance, (multi) physics simulators are mostly implemented in languages such as `FORTRAN`, `C` or `C++`.
 
@@ -67,7 +67,7 @@ But while `OpenGeoSys` is a powerful FEM code, setting up, running and evaluatin
 
 Here is where languages such as `R` and `Python` can prove useful.
 Via an interface that adds a layer on top of `OpenGeoSys 6`, the user can access preprocessing tools, the solver itself and postprocessing tools alike, thus increasing usability and accessibility.
-The development and application of high--level programming and/or scripting languages interfaces to geoscientific simulators   has been gaining increasing attention in recent years. Examples are a) `FloPy` [@Bakker2016]: a `Python` interface to the groundwater simulator `Modflow` [@Langevin2021],  `RedModRPhree` [@DeLucia2021]: `R` utility functions to the geo-chemical solver `Phreeqc` [@Charlton2011], the `toughio` [@Luu2020] Python interface to the multi-phase flow simulator `TOUGH`  [@Pruess2004], the Python interfaces to `OpenGeoSys 5`: `ogs5py` [@Mueller2021] and `OpenGeoSys 6`: `ogs6py` [@Buchwald2021] as well as the `R` interface to `OpenGeoSys 5`: `r2ogs5` [@Schad2023]. 
+The development and application of high--level programming and/or scripting languages interfaces to geoscientific simulators   has been gaining increasing attention in recent years. Examples are `FloPy`: a `Python` interface to the groundwater simulator `Modflow` [@Bakker2016, @Langevin2021],  `RedModRPhree`: an `R` package with utility functions to the geo-chemical solver `Phreeqc` [@DeLucia2021, @Charlton2011], the `toughio` Python interface to the multi-phase flow simulator `TOUGH`  [@Luu2020, @Pruess2004], the Python interfaces to `OpenGeoSys 5`: `ogs5py` [@Mueller2021] and `OpenGeoSys 6`: `ogs6py` [@Buchwald2021] as well as the `R` interface to `OpenGeoSys 5`: `r2ogs5` [@Schad2023]. 
 `FloPy`,  `ogs5py`, `ogs6py` and `r2ogs5` provide object-oriented frameworks to set-up, call the executables and import the output of the associated simulators.
[…]
