+++ JOSS/paper.bib
@@ -87,6 +87,72 @@ address = {Cham}
   journal = {Journal of Open Source Software}
 }
 
+@article{Grunwald2020,
+title = {Non-iterative phase-equilibrium model of the H2O-CO2-NaCl-system for large-scale numerical simulations},
+journal = {Mathematics and Computers in Simulation},
+volume = {178},
+pages = {46-61},
+year = {2020},
+issn = {0378-4754},
+doi = {https://doi.org/10.1016/j.matcom.2020.05.024},
+url = {https://www.sciencedirect.com/science/article/pii/S037847542030183X},
+author = {Norbert Grunwald and Jobst Maßmann and Olaf Kolditz and Thomas Nagel},
+keywords = {Phase equilibrium, CO-HO-NaCl-system, Saline aquifer, CO storage, OpenGeoSys},
+abstract = {This article features a simple method for describing the phase equilibrium of the ternary water–salt–carbon dioxide system. At first, an iterative solution is shown using well known equations of state and solubility correlations to predict the compositions of liquid and gas phases in a saline aquifer. By virtue of several assumptions, iterations can be avoided entirely in order to use the method for demanding large-scale numerical simulations. Subsequently, the iterative method as well as the simplification are evaluated against experimental data. Both solutions provide adequate accuracy for engineering problems such as subsurface carbon dioxide storage.}
+}
+
+@article{Pujades2023,
+  title={Towards the construction of representative regional hydro (geo) logical numerical models: Modelling the upper Danube basin as a starting point},
+  author={Pujades, Estanislao and Kumar, Rohini and Houben, Timo and Jing, Miao and Rakovec, Oldrich and Kalbacher, Thomas and Attinger, Sabine},
+  journal={Frontiers in Earth Science},
+  volume={11},
+  pages={1061420},
+  year={2023},
+  publisher={Frontiers Media SA}
+}
+
+@article{Li2014,
+  title={OpenGeoSys-ChemApp: a coupled simulator for reactive transport in multiphase systems and application to CO 2 storage formation in Northern Germany},
+  author={Li, Dedong and Bauer, Sebastian and Benisch, Katharina and Graupner, Bastian and Beyer, Christof},
+  journal={Acta Geotechnica},
+  volume={9},
+  pages={67--79},
+  year={2014},
+  publisher={Springer}
+}
+
+@article{Jing2018,
+  title={Improved regional-scale groundwater representation by the coupling of the mesoscale Hydrologic Model (mHM v5. 7) to the groundwater model OpenGeoSys (OGS)},
+  author={Jing, Miao and He{\ss}e, Falk and Kumar, Rohini and Wang, Wenqing and Fischer, Thomas and Walther, Marc and Zink, Matthias and Zech, Alraune and Samaniego, Luis and Kolditz, Olaf and others},
+  journal={Geoscientific Model Development},
+  volume={11},
+  number={5},
+  pages={1989--2007},
+  year={2018},
+  publisher={Copernicus GmbH}
+}
+
+@article{Pitz2023,
+  title={Benchmarking a new TH2M implementation in OGS-6 with regard to processes relevant for nuclear waste disposal},
+  author={Pitz, Michael and Grunwald, Norbert and Graupner, Bastian and Radeisen, Eike and Ma{\ss}mann, Jobst and Ziefle, Gesa and Thiedau, Jan and Nagel, Thomas},
+  journal={Environmental Earth Sciences},
+  volume={82},
+  number={13},
+  pages={1--24},
+  year={2023},
+  publisher={Springer}
+}
+
+@article{Asher2015,
+  title={A review of surrogate models and their application to groundwater modeling},
[…]
