### 2023-10-01 #5360 @arfon
@jsta – I've generated a list of possible reviewers by looking at the most similar previous papers over here: https://github.com/openjournals/joss-reviews/issues/5067#issuecomment-1741990601 – if any of these suggestions look good, please approach previous _authors_ before tagging historical reviewers.
