### 2023-12-13 #5360 @kthyng
Hi @ldecicco-USGS and @rchlumsk! Are you interested in reviewing this submission? We had a reviewer disappear and would love to have your expert eyes on this software. We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
