### 2023-12-14 #5360 @waternumbers
@kthyng sorry was in normal journal mode, viewed these as minor corrections so didn't expect to see the changes....
I see @joboog has replied to one set of issues I raised but will presume both sets are fixed are revise review as though they are
