### 2023-12-29 #5360 @danielskatz
Please note the [JOSS submission requirements](https://joss.readthedocs.io/en/latest/submitting.html#submission-requirements) include:
> The software must be hosted at a location where users can open issues and propose code changes without manual approval of (or payment for) accounts.
