### 2023-12-29 #5360 @ldecicco-USGS
Good to know @danielskatz , hopefully that's an easy fix.

My first issue is figuring out where the "your_ogs6_bin_path" is. I installed in my Terminal:
```
python -m venv .venv
pip install ogs
```

I do not see a bin path anywhere in the ogs folder. Are there additional instructions for finding that? I'm on Windows, and I don't have administrative access to my computer. I'll need to file an IT ticket to install OGS if it's more complicated than the pip installation.
