### 2024-01-10 joboog/r2ogs6#1 @ldecicco-USGS
Unable to import *.pvd
I'm having trouble running the simulation. I'm not 100% sure I have everything installed correctly.
First, I don't have access to "conda", so I can't use the installation instructions exactly.

In my Terminal, I ran:
```
pip install numpy
pip install vtk
```

```r
> ogs6_run_simulation(ogs6_obj)
Error in ogs6_read_output_files(ogs6_obj = ogs6_obj) :
  Output file not written out correctly.
                    Unable to import *.pvd
```

Does this message ring a bell that some installation is not done correctly?
