### 2024-01-26 joboog/r2ogs6#1 @joboog
Sorry for the late reply.

It seems that there is a problem with the VTK library, which version did you install?
You really need to install VTK v8.2.0 as shown on the [README](https://gitlab.opengeosys.org/ogs/tools/r2ogs6).
