### 2024-01-26 joboog/r2ogs6#1 @ldecicco-USGS
My organization doesn't allow us to use conda, so I had tried this first:

```
pip install vtk==8.2.0
ERROR: Could not find a version that satisfies the requirement vtk==8.2.0 (from versions: 9.2.4, 9.2.5, 9.2.6, 9.3.0rc1, 9.3.0rc2, 9.3.0, 9.3.20230807rc0)
ERROR: No matching distribution found for vtk==8.2.0
```

Are there other ways to get 8.2.0, I'm only seeing the `conda` instructions? I'm on Windows.
