### 2024-02-05 joboog/r2ogs6#1 @joboog
It seems like VTK abandoned v8.2.0 from [PyPi](https://pypi.org/project/vtk/#history).
I will try to make VTK v9.3.0 work.

Could you try VTK v8.1.2 instead?
