### 2024-02-08 joboog/r2ogs6#1 @ldecicco-USGS
The only versions I can install are:
9.2.4, 9.2.5, 9.2.6, 9.3.0rc1, 9.3.0rc2, 9.3.0, 9.3.20230807rc0

Sorry my computer is locked down so hard, it's been a recent change that we've had and I'm not use to the limitations...
