### 2024-02-08 #5360 @ldecicco-USGS
In case there's a question, we're debugging installation questions here:
https://github.com/joboog/r2ogs6/issues/1
