### 2024-02-12 #5360 @joboog
> I'm taking over as editor!
>
> @ldecicco-USGS Thank you for the update! Yes it's hard to tell exactly where any conversation is occurring...
>
> @joboog Did you migrate to github? If so I can update the repo address in this issue.

@kthyng Yes, the package is available on [github](https://github.com/joboog/r2ogs6) despite not fully migrated yet.
