### 2024-02-19 #5360 @kthyng
> Until the r2ogs6 package can handle vtk 9.3.0 which is the version that is linked from the doc (https://pypi.org/project/vtk/), I don't think I can review the package functionality

@joboog Is this something you can do? Other potential users may have this issue too.
