### 2024-02-25 joboog/r2ogs6#1 @joboog
Hi @ldecicco-USGS ,

I think I misunderstood the cause, sorry.
The problem is not `VTK` it is the actual `OGS` version which is not supported by r2ogs6 yet.
Please uninstall `ogs` and install version 6.4.4.
You can use `vtk==9.3.0`.

```
pip uninstall ogs
pip install ogs==6.4.4
```
