### 2024-02-25 4ac39b9
[config] mv biocon pkgs to imports
