### 2024-03-03 af01bc5
[cfg] add fun to install ogs via python
