### 2024-03-03 f08c664
[test] fun to install ogs
