### 2024-03-03 5306594
[doc] for fun to install ogs
