+++ man/set_ogs6_bin_path.Rd
@@ -1,0 +1,21 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/config_ogs.R
+\name{set_ogs6_bin_path}
+\alias{set_ogs6_bin_path}
+\title{Set default OpenGeoSys 6 binary path}
+\usage{
+set_ogs6_bin_path()
+}
+\value{
+None, but it sets an option `r2ogs6.default_ogs6_bin_path` to point
+to the detected OGS binary path.
+}
+\description{
+This function automatically detects and sets the path to the OpenGeoSys (OGS)
+binary within the active Python virtual environment. It is typically called
+after installing OGS using the `install_ogs` function.
+}
+\examples{
+set_ogs_bin_path()
+
+}
