### 2024-03-03 68cdb08
[doc] update readme to install ogs, numpy, vtk
