### 2024-03-03 314691b
[doc] rm for unused dowload ogs fun
