+++ man/ogs6_download_ogs6.Rd
@@ -1,18 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/download_ogs6.R
-\name{ogs6_download_ogs6}
-\alias{ogs6_download_ogs6}
-\title{Download OpenGeoSys 6}
-\usage{
-ogs6_download_ogs6(path, os, set_as_default_bin_path = TRUE)
-}
-\arguments{
-\item{path}{String: Download path. Defaults to ...}
-
-\item{os}{String: Operating System.}
-
-\item{set_as_default_bin_path}{flag: Defaults to `TRUE`}
-}
-\description{
-Downloads latest prebuilt OpenGeoSys 6 version
-}
