### 2024-03-03 01ca719
[ci] add variable to run test for install_ogs
