+++ .gitlab-ci.yml
@@ -9,7 +9,7 @@ test:
   image: $CI_REGISTRY_IMAGE:24032023
   stage: test
   tags:
-    - envinf2
+    - docker
   variables:
     RETICULATE_MINICONDA_PATH: "/root/.local/share/r-miniconda/"
     RETICULATE_PYTHON_ENV: "/root/.local/share/r-miniconda/envs/r-reticulate/"
