### 2024-03-03 b9ff14c
[base] bugfix example install_ogs
