### 2024-03-03 00bed7b
[ci] switch to plain ubuntu docker image
