### 2024-04-30 c5439dc
[cfg] add checks to set_ogs_bin_path
