### 2024-05-01 eaf9d8c
[ci] use renv restore in test pipeline
