### 2024-05-01 5fd57a2
[cfg] update install_ogs
