### 2024-05-01 d122389
[cfg] update install_ogs
