### 2024-05-01 ea3e916
[cfg] update set_ogs_bin_path
