### 2024-05-01 b997e92
[doc] update install_ogs set_ogs_bin_path
