### 2024-05-02 ae0ef67
[test] bugfix test_config_ogs
