### 2024-05-02 2166bf3
[ci] add test_install job
