### 2024-05-02 joboog/r2ogs6#1 @joboog
@ldecicco-USGS Sorry for the late reply. You are in the wrong directory. The ogs executable is called `ogs.exe` and not located in the module's bin directory. Please check `C:\Users\ldecicco\AppData\Roaming\Python` if there is a `bin` or `Scripts` directory including `ogs.exe`.
