### 2024-05-02 7f2a8e4
[cfg] update install_ogs
