### 2024-05-02 0c20a3b
[doc] update install_ogs
