### 2024-05-02 979950a
[cfg] update Roxygen version in DESCRIPTION
