+++ DESCRIPTION
@@ -48,7 +48,7 @@ Imports:
     foreach,
     BiocManager,
     rhdf5
-RoxygenNote: 7.2.3
+RoxygenNote: 7.3.1
 VignetteBuilder: knitr
 Depends: 
     R (>= 2.10)
