### 2024-05-02 3972744
[cfg] update set_ogs6_bin_path
