+++ R/config_ogs.R
@@ -3,6 +3,7 @@
 
 #' Install specific version of OpenGeoSys 6 (OGS) along with necessary Python 
 #' packages `vtk` and `numpy` into a specified Python virtual environment.
+#' Note, this does not work with `conda` environments.
 #'
 #' @param ogs_version The version of OGS to install (default: 6.4.4).
 #' @param envname The name of the Python virtual environment (default: r2ogs6).
@@ -15,6 +16,17 @@ install_ogs <-
            ogs_version = "6.4.4",
            envname = "r2ogs6"
     ) {
+
+    if (!(reticulate::virtualenv_exists(envname))){
+        stop(
+            paste0('The Python virtual environment "', envname,
+            '" does not exist. You can create it with ',
+            'reticulate::virtualenv_create("r2ogs6").\n',
+            'Make sure that this environment is used as the actual ',
+            'Python configuration; see reticulate::py_config().'
+            )
+        )
+    }
 
     pkgs <- c(paste0("ogs==", ogs_version), "numpy", "vtk")
 
@@ -51,18 +63,25 @@ set_ogs6_bin_path <-
             isTRUE(has_ogs)
         )
 
-        ogs_bin_dir <- file.path(cfg$virtualenv, "bin")
-        files <- list.files(ogs_bin_dir, full.names = TRUE)
+        # Search for ogs binary in potential paths        
         if (Sys.info()["sysname"] == "Windows") {
+            bin_dir <- "Scripts"
             ogs_name <- "ogs.exe"
         } else {
+            bin_dir <- "bin"
             ogs_name <- "ogs"
         }
+        ogs_bin_dir <- file.path(cfg$virtualenv, bin_dir)
+        files <- list.files(ogs_bin_dir, full.names = TRUE)
         ogs_bin_path <- files[grepl(paste0(ogs_name, "$"), basename(files))]
         
-        assertthat::assert_that(file.exists(ogs_bin_path))
         assertthat::assert_that(
-            basename(ogs_bin_path) %in% c("ogs.exe", "ogs")
+            file.exists(ogs_bin_path),
+            msg = "OGS executable not found."
+        )
+        assertthat::assert_that(
+            basename(ogs_bin_path) %in% c("ogs.exe", "ogs"),
+            msg = "OGS executable not found."
         )
         options("r2ogs6.default_ogs6_bin_path" = ogs_bin_path)
 
