+++ README.md
@@ -11,40 +11,10 @@ Furthermore, the package comes with functionality to define [ensemble runs](vign
 
 ## Setup
 
-### Preparing your environment
-
 `r2ogs6` requires an installation of OpenGeoSys 6.
 Furthermore, Python including the libraries `numpy` and `vtk` is required to read in the `.vtu` and `.pvd` files produced by OpenGeoSys 6.
-
-
-You can install OpenGeoSys 6 and Python including `numpy` and `vtk` via a 
-convenient function after having installed `r2ogs6`.
-Just proceed with [Installation]#installation.
-Or do the manual installation.
-
-
-#### Manual Installation of Python and OpenGeoSys 6
-
-Ideally, you do this in a specific virutal python environment. 
-
-```
-# Create python environment
-python -m ogs_env ~/.ogs_env
-source ~/.ogs_env/bin/activate
-
-# Use the ogs version appropriate for the r2ogs6 version, see releases
-pip install ogs==6.4.4
-
-# Install python dependencies
-pip install numpy vtk
-```
-
-Alternatively, can download OpenGeoSys 6 executables [here](https://www.opengeosys.org/releases/) and set up the python 
-environment with `numpy` and `vtk` only.
-For more information on how to install OpenGeoSys 6 look [here](https://www.opengeosys.org/docs/userguide/basics/introduction/#install-via-pip).
-Be aware, to download or install the OpenGeoSys 6 version compatible with the `r2ogs6` version, see [releases](https://gitlab.opengeosys.org/ogs/tools/r2ogs6/-/releases).
+These can be install after having installed `r2ogs6`.
 `r2ogs6` was tested with the following environment: `python==3.10.12`, `vtk==9.3.0`, `numpy==1.26.4`
-If you don't have any Python installation on your system, you can download the latest release of Python from [here](https://www.python.org/downloads/). 
 
 
 ### r2ogs6 Installation
@@ -63,13 +33,13 @@ To install `r2ogs6`, you first need the R-packages `remotes` and `BiocManager`.
 `remotes` is used for installing `r2ogs6` and `BiocManager` to set the URLs 
 to search dependencies on *CRAN* and *bioconductor.org*.
 For instance, the dependency `rhdf5` is only available on *bioconductor.org*.
-In your R console:
+Open a R console in your specific project:
 
 ```r
 install.packages(c("remotes", "BiocManager"))
 ```
 
-With `remotes`, you can now install `r2ogs6`.
+With `remotes`, you can now install `r2ogs6` in your preferred library.
 
 ```r
 # Install r2ogs6. Change the path to that of the cloned repository!
@@ -80,15 +50,109 @@ remotes::install_local(
 )
 ```
 
-Now install OpenGeoSys 6 and Python including `numpy` and `vtk`.
+#### Set-up Python
+
+OpenGeoSys 6 requires a Python installation of version 3.8--3.11.
+
+
+##### Linux systems
+
+If you do not have Python installed yet you can do so with the 
+[reticulate](https://rstudio.github.io/reticulate/index.html) library from
+your R console.
+
+```r
+reticulate::install_python(version = "3.10:latest")
+```
+
+Otherwise you can use the package manager of your linux distribution.
+
+
+##### Windows
+
+Ideally, install Python with a [Python Windows installer](https://www.python.org/downloads/windows/).
+Of course you could also try:
+
+```r
+reticulate::install_python(version = "3.10:latest")
+```
+
+#### Install Python dependencies and OpenGeoSys 6
+
+Create a Python virtual environment with a specific name (e.g. "r2ogs6) and the
+ path to the Python executable you just installed.
+
+```r
[…]
