### 2024-05-03 6e51f09
[cfg] update set_ogs_path win dirs
