### 2024-05-03 2846033
[test] update config ogs to win path
