### 2024-05-03 334acb3
[cfg] add fun to install ogs python module
