### 2024-05-03 cae56d4
[cfg] add fun to set ogs6 bin path as option
