+++ R/config_ogs.R
@@ -38,3 +38,60 @@ install_ogs <-
   }
 
 
+#' Set default OpenGeoSys 6 binary path
+#'
+#' This function automatically detects and temporarily sets the path to the
+#' OpenGeoSys (OGS) binary within the active Python virtual environment. It is
+#' typically called after installing OGS using the `install_ogs()` function.
+#' For a permanent setting of the OGS binary path, consider to define
+#' `r2ogs6.default_ogs6_bin_path` in a `config.yml` file.'
+#'
+#' @return None, but it sets an option `r2ogs6.default_ogs6_bin_path` to point
+#' to the detected OGS binary path.
+#'
+#' @export
+set_ogs6_bin_path <-
+    function() {
+        cfg <- reticulate::py_config()
+
+        has_python <- reticulate::py_available(initialize = FALSE)
+        has_numpy <- reticulate::py_numpy_available(initialize = FALSE)
+        has_vtk <- reticulate::py_module_available("vtk")
+        has_ogs <- reticulate::py_module_available("ogs")
+        assertthat::assert_that(
+            isTRUE(has_python), isTRUE(has_numpy), isTRUE(has_vtk),
+            isTRUE(has_ogs)
+        )
+
+        # Search for ogs binary in potential paths
+        if (Sys.info()["sysname"] == "Windows") {
+            bin_dir <- "Scripts"
+            ogs_name <- "ogs.exe"
+        } else {
+            bin_dir <- "bin"
+            ogs_name <- "ogs"
+        }
+        virtualenv_path <- win_to_linux_path(cfg$virtualenv)
+        ogs_bin_dir <- file.path(virtualenv_path, bin_dir)
+        files <- list.files(ogs_bin_dir, full.names = TRUE)
+        ogs_bin_path <- files[grepl(paste0(ogs_name, "$"), basename(files))]
+
+        assertthat::assert_that(
+            file.exists(ogs_bin_path),
+            msg = "OGS executable not found."
+        )
+        assertthat::assert_that(
+            basename(ogs_bin_path) %in% c("ogs.exe", "ogs"),
+            msg = "OGS executable not found."
+        )
+        options("r2ogs6.default_ogs6_bin_path" = ogs_bin_path)
+
+        message(
+            paste0(
+                'The option "r2ogs6.default_ogs6_bin_path" was temporarily ',
+                'set to "', ogs_bin_path, '" .\n',
+                'For a permanent setting of "r2ogs6.default_ogs6_bin_path=',
+                ogs_bin_path, ' define it in a "config.yml" file.'
+            )
+        )
+  }
