### 2024-05-03 e00c43f
[cfg] add install_ogs to  NAMESPACE
