+++ NAMESPACE
@@ -8,8 +8,8 @@ export(OGS6_h5)
 export(OGS6_msh)
 export(OGS6_pvd)
 export(OGS6_vtu)
+export(install_ogs)
 export(ogs6_call_ogs6)
-export(ogs6_download_ogs6)
 export(ogs6_export_sim_files)
 export(ogs6_generate_benchmark_script)
 export(ogs6_generate_benchmark_scripts)
@@ -71,6 +71,7 @@ export(prj_time_stepping)
 export(prj_tl_process)
 export(prj_vtkdiff)
 export(read_in_prj)
+export(set_ogs6_bin_path)
 importFrom(R6,R6Class)
 importFrom(crayon,green)
 importFrom(crayon,red)
