+++ NAMESPACE
@@ -80,3 +80,5 @@ importFrom(doParallel,registerDoParallel)
 importFrom(dplyr,bind_cols)
 importFrom(dplyr,bind_rows)
 importFrom(foreach,"%dopar%")
+importFrom(rhdf5,h5ls)
+importFrom(rhdf5,h5read)
