### 2024-05-03 1cbbfe7
[cfg] update  dependencies to newest versions
