### 2024-05-03 535033b
[ci] update testing job
