+++ .gitlab-ci.yml
@@ -5,25 +5,44 @@
 stages:
   - test
 
+variables:
+    # RETICULATE_MINICONDA_PATH: "/root/.local/share/r-miniconda/"
+    # RETICULATE_PYTHON_ENV: "/root/.local/share/r-miniconda/envs/r-reticulate/"
+    RENV_CONFIG_EXTERNAL_LIBRARIES: "/usr/local/lib/R/library"
+    RENV_PATHS_CACHE: "/root/.cache/R/renv/cache" #"$CI_PROJECT_DIR/cache"
+    RENV_PATHS_LIBRARY: "/root/.cache/R/renv/library" #"$CI_PROJECT_DIR/renv/library"
+    RUN_PYTHON_CONFIG_TESTS: "false"
+    R_LIBS: "/usr/local/lib/R/site-library:/usr/local/lib/R/library:$RENV_PATHS_LIBRARY"
+
+.before_script_templ: &before_script_templ
+  - apt-get update && apt-get upgrade -y
+  - apt-get install -y git
+  # Packages to build Python and R packages
+  - |
+    apt-get install -y build-essential libssl-dev zlib1g-dev libbz2-dev \
+    libreadline-dev libsqlite3-dev curl libncursesw5-dev xz-utils tk-dev \
+    libxml2-dev libxmlsec1-dev libffi-dev liblzma-dev \
+    libharfbuzz-dev libfribidi-dev libcurl4-openssl-dev \
+    libfreetype6-dev libpng-dev libtiff5-dev libjpeg-dev
+  - apt-get install -y python3.10-venv python3-pip
+
 test:
-  image: $CI_REGISTRY_IMAGE:24032023
+  image: rocker/r-ver:4.2.3
   stage: test
   tags:
-    - envinf2
-  variables:
-    RETICULATE_MINICONDA_PATH: "/root/.local/share/r-miniconda/"
-    RETICULATE_PYTHON_ENV: "/root/.local/share/r-miniconda/envs/r-reticulate/"
-    RENV_CONFIG_EXTERNAL_LIBRARIES: "/usr/local/lib/R/library"
-    RENV_PATHS_CACHE: "$CI_PROJECT_DIR/cache"
-    RENV_PATHS_LIBRARY: "$CI_PROJECT_DIR/renv/library"
+    - docker
   before_script:
-    - R -e "if (!requireNamespace('renv', quietly = TRUE)) install.packages('renv')"
-    #- R -e 'renv::restore()' does fail while loading Bioconductor packages
-    - R -e 'renv::install()'
-    - R -e 'renv::install("bioc::rhdf5")'
+    - *before_script_templ
+    - R -e "install.packages('renv', repos = c(CRAN = 'https://cloud.r-project.org'))"
+    - R -e "renv::restore()"
+    #- R -e 'renv::status()'
+    # - R -e 'renv::install()'
+    # - R -e 'renv::install("bioc::rhdf5")'
   script:
-    - R CMD build . --no-manual
-    - R CMD check $(ls -1t *.tar.gz | head -n 1) --no-manual
+    - R -e "devtools::test()"
+    - R -e "devtools::check(vignettes=FALSE)"
+    # - R CMD build . --no-manual --no-build-vignettes
+    # - R CMD check $(ls -1t *.tar.gz | head -n 1) --no-manual --no-build-vignettes
   after_script:
     - R -e 'covr::package_coverage(type = c("tests", "examples"), quiet = F)'
   coverage: '/Coverage: \d+\.\d+\%/'
