### 2024-05-03 4f3265d
[ci] add job to test installation on linux
