### 2024-05-03 3b5ab46
[doc] update Readme to new install_ogs fun
