+++ README.md
@@ -11,55 +11,148 @@ Furthermore, the package comes with functionality to define [ensemble runs](vign
 
 ## Setup
 
-### Preparing your environment
-
-Before you install and load `r2ogs6`, there are two prerequisites that need to be met so you can use `r2ogs6` to its full extent later.
-
-The most important one is having OpenGeoSys 6 installed on your system. You can download the current version of OpenGeoSys 6 from [here](https://www.opengeosys.org/releases/). 
-
-Secondly, to read in the `.vtu` and `.pvd` files produced by OpenGeoSys 6, you need a Python installation including the libraries `numpy` and `vtk`. If you have an Anaconda installation on your system, setting up a Python environment with `numpy` and `vtk` will work. `r2ogs6` was tested with the following environment:
-
-```
-$ conda create -n r2ogs6 python=3.7
-$ conda install -c anaconda numpy, vtk==8.2.0
-```
-
-If you don't have any Python installation on your system, you can download the latest release of Python from [here](https://www.python.org/downloads/). Instructions on how to install `numpy` and `vtk` can be found [here](https://numpy.org/install/) and  [here](https://pypi.org/project/vtk/).
+`r2ogs6` requires an installation of OpenGeoSys 6.
+Furthermore, Python including the libraries `numpy` and `vtk` is required to read in the `.vtu` and `.pvd` files produced by OpenGeoSys 6.
+These can be install after having installed `r2ogs6`.
+`r2ogs6` was tested with the following environment: `python==3.10.12`, `vtk==9.3.0`, `numpy==1.26.4`
 
 
-### Installation
+### r2ogs6 Installation
 
 First, open a terminal and clone the `r2ogs6` repository to your local machine.
 
 ```
-$ git clone https://gitlab.opengeosys.org/ogs/tools/r2ogs6.git
+git clone https://gitlab.opengeosys.org/ogs/tools/r2ogs6.git
+cd r2ogs6
+
+# Now checkout the latest stable release
+git checkout tags/v0.4.643
 ```
 
-After that, open R from your local (anaconda) environment.
-To install `r2ogs6`, you first need the R-package `devtools`.
+To install `r2ogs6`, you first need the R-packages `remotes` and `BiocManager`.
+`remotes` is used for installing `r2ogs6` and `BiocManager` to set the URLs 
+to search dependencies on *CRAN* and *bioconductor.org*.
+For instance, the dependency `rhdf5` is only available on *bioconductor.org*.
+Open a R console in your specific project:
 
 ```r
-# Install devtools
-install.packages("devtools")
+install.packages(c("remotes", "BiocManager"))
 ```
 
-With `devtools`, you can now install `r2ogs6`.
+With `remotes`, you can now install `r2ogs6` in your preferred library.
 
 ```r
 # Install r2ogs6. Change the path to that of the cloned repository!
-devtools::install("path/to/r2ogs6")
-
-# Alternatively, install directly from the Gitlab repo:
-devtools::install_git("https://gitlab.opengeosys.org/ogs/tools/r2ogs6")
+remotes::install_local(
+    path="path/to/r2ogs6", 
+    dependencies="Imports",
+    repos=BiocManager::repositories()
+)
 ```
 
-This will install all necessary R dependencies except `rhdf5`.  
-This package is needed to handle *hdf5* output files and can be installed via:
+#### Set-up Python
+
+OpenGeoSys 6 requires a Python installation of version 3.8--3.11.
+
+
+##### Linux systems
+
+If you do not have Python installed yet you can do so with the 
+[reticulate](https://rstudio.github.io/reticulate/index.html) library from
+your R console.
+
 ```r
[…]
