### 2024-05-03 372f545
[JOSS] clean up final paragraph in paper
