### 2024-05-03 5516983
[JOSS] update links to package web articles
