### 2024-05-03 #5360 @joboog
> # Package feedback
> At this point I'm still unable to install the proper version of vtk (8.2). Until the `r2ogs6` package can handle vtk 9.3.0 which is the version that is linked from the doc (https://pypi.org/project/vtk/), I don't think I can review the package functionality. All of the functions that are used to create the ogs6_obj work perfectly. I therefore assume `ogs6_run_simulation` and `ogs6_read_output_files` would work - but they require the older vtk library.
>
> Since this package is for R users, it will be important to include detailed instructions on how to install the mandatory Python libraries. The reason I would choose this package over the Python ogs6py package is because I'm much more comfortable in R. The reason I'm more comfortable in R is because installing things in Python gives me a headache 😂
>
> ##########
>
> # Paper feedback:
> The paper nicely outlines why someone would be interested in this package, and gives a nice example of how to use it. I would recommend cleaning up the final paragraph.
>
> "Please, check the following package vignettes for more information: a) a further guide on how to create ensemble runs ([link](https://gitlab.opengeosys.org/ogs/tools/r2ogs6/-/blob/master/vignettes/ensemble_workflow_vignette.Rmd)), b) a tutorial to set up a single simulation of a hydro-mechanics benchmark with the package functions ([link](https://gitlab.opengeosys.org/ogs/tools/r2ogs6/-/blob/master/vignettes/user_workflow_vignette.Rmd)), and c) a guide how to start to further develop the package ([link](https://gitlab.opengeosys.org/ogs/tools/r2ogs6/-/blob/master/vignettes/dev_workflow_vignette.Rmd))."
>
> Would it be possible to create a GitLab page using `pkgdown`, and link to a rendered version of the Rmd? I'd also suggest removing the "Please, " to just read:
>
> "Check the following package vignettes...."

I updated the package and the paper accordingly.

**@Package Feedback**
- Added support to use VTK v9.0.3
- There is now also a function to install OpenGeoSys 6 as python module (`install_ogs()`)
- The section [r2ogs6-installation](https://github.com/joboog/r2ogs6?tab=readme-ov-file#r2ogs6-installation) was revised accordingly.

**@Paper Feedback**
- The paragraph was revised to

  "Check the following package vignettes for more information:
  - a guide to set up a single simulation of a hydro-mechanics benchmark ([link](https://joboog.github.io/r2ogs6/articles/user_workflow_vignette.html))
  - a guide to create ensemble runs ([link](https://joboog.github.io/r2ogs6/articles/ensemble_workflow_vignette.html))
  - a development guide ([link](https://joboog.github.io/r2ogs6/articles/dev_workflow_vignette.html))"

- Gitlab page with vignettes was created [link](https://joboog.github.io/r2ogs6)
