### 2024-05-03 joboog/r2ogs6#1 @joboog
The master branch contains an updated version including a more handy way to install the Python environment and function to install OpenGeoSys. Please see the [README](https://joboog.github.io/r2ogs6/index.html). I tested it on Windows 10 with Python 3.11.
