+++ _pkgdown.yml
@@ -1,0 +1,4 @@
+url: https://joboog.github.io/r2ogs6
+template:
+  bootstrap: 5
+
