### 2024-06-03 #5360 @ldecicco-USGS
The `reticulate` instructions are great! As an R user, that seemed a lot more streamlined for setting up the environments.

That being said, I'm still seeing:

```
Error in ogs6_read_output_files(ogs6_obj = ogs6_obj) :
  Output file not written out correctly.
                    Unable to import *.pvd
```

When I look at the `r2ogs6` CI tests, the pipeline is failing in the same way:

https://gitlab.opengeosys.org/ogs/tools/r2ogs6/-/jobs/434727
