### 2024-06-08 8ba76a3
[ci] def pkgs when installing python env
