### 2024-06-09 5ff9118
[doc] add manuals for config ogs funs
