### 2024-06-09 8682e2c
[ci] set env vars for specific job
