### 2024-06-09 00e30f0
[ci] run example in job test_install
