+++ .gitlab-ci.yml
@@ -68,6 +68,7 @@ test_install:
     - *before_script_templ
     - mkdir ../test_r2ogs6 && cd ../test_r2ogs6 && ls . -ahl
   script:
+    # Install r2ogs6
     - R -e 'install.packages(c("remotes", "BiocManager"))'
     - R -e "remotes::install_local(path=\"$CI_PROJECT_DIR\", dependencies='Imports', repos=BiocManager::repositories())"
     - R -e 'reticulate::virtualenv_create(envname="r2ogs6", python="/usr/bin/python3", packages = c("numpy", "vtk"))'
@@ -75,6 +76,16 @@ test_install:
     - R -e 'reticulate::py_config()'
     - R -e 'r2ogs6::install_ogs(ogs_version = "6.4.4", envname = "r2ogs6")'
     - R -e 'reticulate::py_module_available("ogs")'
+    # Run benchmark
+    - |
+      R -e '
+      library(r2ogs6)
+      set_ogs6_bin_path();
+      prj_path <- system.file("extdata/benchmarks/flow_no_strain/flow_no_strain.prj", package = "r2ogs6");
+      sim_path <- tempdir();
+      exit_code <- r2ogs6:::run_benchmark(prj_path=prj_path, sim_path=sim_path);
+      assertthat::are_equal(exit_code, 0)
+      '
   rules:
     - if: '$CI_PIPELINE_SOURCE == "merge_request_event" || $CI_PIPELINE_SOURCE == "web"'
     - if: '$CI_COMMIT_BRANCH == "master"'
