### 2024-06-09 2c80e7b
[doc] update setup in README
