### 2024-06-09 b15deec
[doc] add example in README
