### 2024-06-10 1390315
[ci] set python path through renviron in test_install job
