+++ .gitlab-ci.yml
@@ -72,7 +72,8 @@ test_install:
     - R -e 'install.packages(c("remotes", "BiocManager"))'
     - R -e "remotes::install_local(path=\"$CI_PROJECT_DIR\", dependencies='Imports', repos=BiocManager::repositories())"
     - R -e 'reticulate::virtualenv_create(envname="r2ogs6", python="/usr/bin/python3", packages = c("numpy", "vtk"))'
-    - R -e 'reticulate::use_virtualenv(virtualenv="r2ogs6")'
+    #- R -e 'reticulate::use_virtualenv(virtualenv="r2ogs6")'
+    - echo "RETICULATE_PYTHON=/root/.virtualenvs/r2ogs6/bin/python" > .Renviron
     - R -e 'reticulate::py_config()'
     - R -e 'r2ogs6::install_ogs(ogs_version = "6.4.4", envname = "r2ogs6")'
     - R -e 'reticulate::py_module_available("ogs")'
