### 2024-06-10 83bceb6
[ci] explicitly setup python env in test job
