### 2024-06-12 #5360 @joboog
@ldecicco-USGS sometimes `reticulate`  is not pointing to the created virtual python environment where the `ogs` package is installed. Then the simulations are actually not run.
I did some bugfixes (on master branch).
Please try again using the updated instructions in the README of the actual master branch.
