### 2024-08-09 #5360 @ldecicco-USGS
IT WORKED! 🎉
The installation on a Windows computer using reticulate worked following the instructions on the README. Thank you for adding those! As a Windows R user with limited admin access to my own computer (which...is not uncommon for a government worker, or some industries), that is not uncommon.

The response to the original review sounds great!
