### 2024-08-16 #5360 @kthyng
Woohoo!!! Thank you SO MUCH @ldecicco-USGS! You really saved the day... or year as the case may be.

@joboog I went back and found these issues opened by reviewer @waternumbers https://gitlab.opengeosys.org/ogs/tools/r2ogs6/-/issues/86, https://gitlab.opengeosys.org/ogs/tools/r2ogs6/-/issues/87. The first looks like you addressed it, but not the second. Can you update me on the status?
