### 2024-08-25 dda0da9
[doc] add statement of need to README
