### 2024-08-26 #5360 @joboog
@kthyng my replies to https://gitlab.opengeosys.org/ogs/tools/r2ogs6/-/issues/86 had all been given, I am actually working on one last item from https://gitlab.opengeosys.org/ogs/tools/r2ogs6/-/issues/87.
