### 2024-09-01 1eec725
[base] add top level tag raster as class
