### 2024-09-01 932ad72
[base] add child tag to prj_process
