### 2024-09-01 ec69dbc
[base] add child tag to prj_time_loop
