+++ R/prj_time_loop.R
@@ -5,13 +5,15 @@
 #' prj_time_loop
 #' @description tag: time_loop
 #' @param processes list, prj_tl_process:
-#' @param output prj_output:
+#' @param output Optional: prj_output:
+#' @param outputs Optional: list, prj_output:
 #' @param global_process_coupling Optional: prj_global_process_coupling:
 #' @param submesh_residuum_output Optional: prj_submesh_residuum_output:
 #' @example man/examples/ex_prj_time_loop.R
 #' @export
 prj_time_loop <- function(processes,
-                          output,
+                          output = NULL,
+                          outputs = NULL,
                           global_process_coupling = NULL,
                           submesh_residuum_output = NULL) {
 
@@ -20,18 +22,27 @@ prj_time_loop <- function(processes,
 
     new_prj_time_loop(processes,
                         output,
+                        outputs,
                         global_process_coupling,
                         submesh_residuum_output)
 }
 
 
 new_prj_time_loop <- function(processes,
-                                output,
+                                output = NULL,
+                                outputs = NULL,
                                 global_process_coupling = NULL,
                                 submesh_residuum_output = NULL) {
 
     is_wrapper_list(processes, "prj_tl_process")
-    assertthat::assert_that(class(output) == "prj_output")
+    
+    # either output or outputs has to be given
+    assertthat::assert_that((!is.null(output)) | (!is.null(outputs)))
+    assertthat::assert_that(!(!is.null(output) & !is.null(outputs))) 
+    is_null_or_has_class(output, "prj_output")
+    if(!is.null(outputs)){
+        is_wrapper_list(outputs, "prj_output")
+    }
 
     is_null_or_has_class(global_process_coupling,
                                   "prj_global_process_coupling")
@@ -42,6 +53,7 @@ new_prj_time_loop <- function(processes,
         list(
             processes = processes,
             output = output,
+            outputs = outputs,
             global_process_coupling = global_process_coupling,
             submesh_residuum_output = submesh_residuum_output,
             xpath = "time_loop",
@@ -150,6 +162,7 @@ new_prj_tl_process <- function(ref,
 #' @param fixed_output_times Optional: string | numeric:
 #' @param hdf Optional: numeric
 #' @param geometrical_sets Optional: list, prj_geometrical_set
+#' @param output_extrapolation_residuals Optional: string: Either "true" or "false"
 #' @example man/examples/ex_prj_output.R
 #' @export
 prj_output <- function(type,
@@ -163,7 +176,9 @@ prj_output <- function(type,
                           meshes = NULL,
                           fixed_output_times = NULL,
                           hdf = NULL,
-                          geometrical_sets = NULL) {
+                          geometrical_sets = NULL,
+                          output_extrapolation_residuals = NULL
+                ) {
 
     #Coerce input
     fixed_output_times <- coerce_string_to_numeric(fixed_output_times)
@@ -179,7 +194,8 @@ prj_output <- function(type,
                          meshes,
                          fixed_output_times,
                          hdf,
-                         geometrical_sets)
+                         geometrical_sets,
+                         output_extrapolation_residuals)
 }
 
 
@@ -194,7 +210,8 @@ new_prj_output <- function(type,
                                  meshes = NULL,
                                  fixed_output_times = NULL,
                                  hdf = NULL,
-                                 geometrical_sets = NULL) {
+                                 geometrical_sets = NULL,
+                                 output_extrapolation_residuals = NULL) {
 
     assertthat::assert_that(assertthat::is.string(type))
     assertthat::assert_that(assertthat::is.string(prefix))
@@ -202,7 +219,8 @@ new_prj_output <- function(type,
     names(variables) <- rep("variable", length(variables))
 
[…]
