### 2024-09-01 dde6045
[doc] add protocol for updating to ogs6.5.2
