+++ update_protocols/protocol_update_v6.5.2_keywords.R
@@ -1,0 +1,286 @@
+# !!!!! READ the developer guide before updateing r2ogs6 !!!!!!
+# https://gitlab.opengeosys.org/ogs/tools/r2ogs6/-/blob/master/vignettes/dev_workflow_vignette.Rmd
+
+library(dplyr)
+devtools::load_all(".") # load current r2ogs6 version
+
+temppath <- "tmp/update_652"
+
+
+# top level tags -----------------------------------------------------------
+# Both the old and new BM folders are supposed to be stored in the directory
+# now check for new top level tags
+old <- analyse_xml(path = paste0(temppath,"/Data_644"),
+                   pattern = "\\.prj$",
+                   xpath = "*",
+                   print_findings = TRUE)
+
+new <- analyse_xml(path = paste0(temppath,"/Data_652"),
+                    pattern = "\\.prj$",
+                    xpath = "*",
+                    print_findings = TRUE)
+
+ind <- names(new$both_sorted) %in% names(old$both_sorted)
+names(new$both_sorted)[!ind]
+# output: "outputs" "use_high_precision" "raster"
+# "raster" is a new top-level tag
+
+# "use_high_precision" is a child tag of chemical_system, already implemented
+# "outputs" is a child tag ot "time_loop" and a parent wrapper tag of
+# "time_loop/output"
+
+
+
+
+# add top-level tag "rasters"
+analyse_xml(path = paste0(temppath,"/Data_652"),
+                    pattern = "\\.prj$",
+                    xpath = "//rasters",
+                    print_findings = TRUE)
+
+raster <- analyse_xml(path = paste0(temppath,"/Data_652"),
+                    pattern = "\\.prj$",
+                    xpath = "/OpenGeoSysProject/rasters/raster",
+                    print_findings = TRUE)
+
+generate_constructor(raster, print_result = TRUE)
+generate_helper(raster, print_result = TRUE)
+
+# new class prj_raster added
+
+# execute data_raw/xpaths_for_classes.R
+xpaths_for_classes[["prj_raster"]]
+
+# add test with bm Parabolic/LiquidFlow/RasterParameter/GroundwaterModelWithRasterBC.prj
+
+
+
+
+
+
+
+
+
+# top-level/child tags-------------------------------------------------------
+keywords_lvl1 <- names(new$both_sorted)
+change_df <- tibble::tibble(parent = character(), child = character(), new = logical())
+
+# walk over every old level1 tag and check which child elements are found for
+# old and new benchmarks
+for (i in seq_along(keywords_lvl1)) {
+    old <- analyse_xml(path = paste0(temppath,"/Data_644"),
+                        pattern = "\\.prj$",
+                        xpath = paste0("//", keywords_lvl1[i]),
+                        print_findings = FALSE)$both_sorted %>% names
+
+    new <- analyse_xml(path = paste0(temppath,"/Data_652"),
+                         pattern = "\\.prj$",
+                         xpath = paste0("//", keywords_lvl1[i]),
+                         print_findings = FALSE)$both_sorted %>% names
+
+    if (any(!new %in% old)) {
+        change_df <- rbind(change_df,
+                           tibble::tibble(parent = keywords_lvl1[i],
+                                          child = new[!new %in% old],
+                                          new = TRUE))
+        cat("new child: ")
+        cat(new[!new %in% old])
+        cat("\nof tag: ")
+        cat(keywords_lvl1[i])
+        cat("\n")
+        cat(rep("#", 10))
+        cat("\n")
+    }
+    if (any(!old %in% new)) {
+        change_df <- rbind(change_df,
+                           tibble::tibble(parent = keywords_lvl1[i],
[…]
