### 2024-10-27 bde6725
[bugfix] variables in prj_output
