### 2024-10-27 0898983
[base] add new vars to prj_pr_property
