### 2024-10-27 b8e614c
[base] add phase types
