### 2024-10-27 731acc3
[bugfix] b_bar validation
