### 2024-10-27 932a977
[base] add var to prj_ice_constitutive_relation
