### 2024-10-27 c904ee6
[bugfix] abspath
