### 2024-10-27 a76e23a
[base] add class prj_output_mesh
