### 2024-10-27 181d736
[base] specify reading param nodes of zero length
