+++ R/read_in_utils.R
@@ -116,19 +116,36 @@ node_to_prj_class_object <- function(xml_node,
     parameter_nodes <- xml2::xml_children(xml_node)
     parameters <- c(list(), xml2::xml_attrs(xml_node))
 
-    for(i in seq_len(length(parameter_nodes))){
+    if(length(parameter_nodes)==0){
+        xml_text <- xml2::xml_text(xml_node)
+        xml_text_clean <- stringr::str_trim(xml_text)
+        xml_text_clean <-
+            stringr::str_remove_all(xml_text_clean, "[\n]")
+            if(xml_text_clean != ""){
+                parameters <- c(parameters, xml_text_clean)
+            }
+        if(length(parameters)==1){
+            names(parameters) <- "name"
+        }else{
+            names(parameters)[2] <- "name"
+        }
+        
+        
+    }else{
+        for(i in seq_len(length(parameter_nodes))){
 
-        new_xpath <- paste0(xpath,
-                             "/",
-                             xml2::xml_name(parameter_nodes[[i]]))
+            new_xpath <- paste0(xpath,
+                                "/",
+                                xml2::xml_name(parameter_nodes[[i]]))
 
-        #Guess R representation of node, add it to parameter list
-        parameters <- c(parameters, list(node_to_object(parameter_nodes[[i]],
-                                                        new_xpath)))
+            #Guess R representation of node, add it to parameter list
+            parameters <- c(parameters, list(node_to_object(parameter_nodes[[i]],
+                                                            new_xpath)))
 
-        #Name parameter after the xml_node child name
-        names(parameters)[[length(parameters)]] <-
-            xml2::xml_name(parameter_nodes[[i]])
+            #Name parameter after the xml_node child name
+            names(parameters)[[length(parameters)]] <-
+                xml2::xml_name(parameter_nodes[[i]])
+    }
     }
 
     class_name <- get_class_from_xpath(xpath)
