### 2024-10-27 046c1dd
[base] option to remove tag name on export
