### 2024-11-03 ae5fa42
[test] update test cfg to ogs 6.5.2
