### 2024-11-03 25db5e3
[test] update test of read prj to prj_mesh
