+++ R/prj_process.R
@@ -666,11 +666,17 @@ new_prj_process <- function(name,
 #' @param type string:
 #' @param id Optional:
 #' @param youngs_modulus Optional:
+#' @param youngs_modulus_i Optional:
+#' @param youngs_modulus_a Optional:
 #' @param poissons_ratio Optional:
+#' @param poissons_ratio_ii Optional:
+#' @param poissons_ratio_ia Optional:
 #' @param nonlinear_solver Optional:
 #' @param behaviour Optional:
 #' @param material_properties Optional:
 #' @param shear_modulus Optional:
+#' @param shear_modulus_ii Optional:
+#' @param shear_modulus_ia Optional:
 #' @param bulk_modulus Optional:
 #' @param kappa Optional:
 #' @param beta Optional:
@@ -708,11 +714,17 @@ new_prj_process <- function(name,
 prj_constitutive_relation <- function(type,
                                          id = NULL,
                                          youngs_modulus = NULL,
+                                         youngs_modulus_i = NULL,
+                                         youngs_modulus_a = NULL,
                                          poissons_ratio = NULL,
+                                         poissons_ratio_ii = NULL,
+                                         poissons_ratio_ia = NULL,
                                          nonlinear_solver = NULL,
                                          behaviour = NULL,
                                          material_properties = NULL,
                                          shear_modulus = NULL,
+                                         shear_modulus_ii = NULL,
+                                         shear_modulus_ia = NULL,
                                          bulk_modulus = NULL,
                                          kappa = NULL,
                                          beta = NULL,
@@ -752,11 +764,17 @@ prj_constitutive_relation <- function(type,
     new_prj_constitutive_relation(type,
                                      id,
                                      youngs_modulus,
+                                     youngs_modulus_i,
+                                     youngs_modulus_a,
                                      poissons_ratio,
+                                     poissons_ratio_ii,
+                                     poissons_ratio_ia,
                                      nonlinear_solver,
                                      behaviour,
                                      material_properties,
                                      shear_modulus,
+                                     shear_modulus_ii,
+                                     shear_modulus_ia,
                                      bulk_modulus,
                                      kappa,
                                      beta,
@@ -795,11 +813,17 @@ prj_constitutive_relation <- function(type,
 new_prj_constitutive_relation <- function(type,
                                              id = NULL,
                                              youngs_modulus = NULL,
+                                             youngs_modulus_i = NULL,
+                                             youngs_modulus_a = NULL,
                                              poissons_ratio = NULL,
+                                             poissons_ratio_ii = NULL,
+                                             poissons_ratio_ia = NULL,
                                              nonlinear_solver = NULL,
                                              behaviour = NULL,
                                              material_properties = NULL,
                                              shear_modulus = NULL,
+                                             shear_modulus_ii = NULL,
+                                             shear_modulus_ia = NULL,
                                              bulk_modulus = NULL,
                                              kappa = NULL,
                                              beta = NULL,
@@ -835,11 +859,17 @@ new_prj_constitutive_relation <- function(type,
     structure(list(type = type,
                    id = id,
                    youngs_modulus = youngs_modulus,
+                   youngs_modulus_i = youngs_modulus_i,
+                   youngs_modulus_a = youngs_modulus_a,
                    poissons_ratio = poissons_ratio,
+                   poissons_ratio_ii = poissons_ratio_ii,
+                   poissons_ratio_ia = poissons_ratio_ia,
                    nonlinear_solver = nonlinear_solver,
                    behaviour = behaviour,
                    material_properties = material_properties,
                    shear_modulus = shear_modulus,
+                   shear_modulus_ii = shear_modulus_ii,
+                   shear_modulus_ia = shear_modulus_ia,
                    bulk_modulus = bulk_modulus,
                    kappa = kappa,
                    beta = beta,
