### 2024-11-03 e2088e7
[base] rm dimension param of prj_process
