+++ man/examples/ex_prj_local_coordinate_system.R
@@ -1,3 +1,7 @@
-prj_local_coordinate_system(basis_vector_0 = "e0",
-                               basis_vector_1 = "e1",
-                               basis_vector_2 = "e2")
+e0 <- prj_basis_vector_0(name = "e0")
+e1 <- prj_basis_vector_1(name = "e1")
+e2 <- prj_basis_vector_2(implicit = "true")
+
+prj_local_coordinate_system(basis_vector_0 = e0,
+                            basis_vector_1 = e1,
+                            basis_vector_2 = e2)
