### 2024-11-17 66639f7
[bugfix] initialize python env when checking for numpy
