### 2024-11-17 04587a7
[bugfix] add variable parameter even when variable are empty
