### 2024-11-17 5598c4b
[bugfix] doc string in prj_output_mesh
