### 2024-11-17 d331ab8
[base] add dummy name param for tags with attributes
