### 2024-11-17 b823012
[doc] add  params b_bar to prj_process
