### 2024-11-17 3168bbb
[doc] add prj_raster to ogs6 class
