+++ NAMESPACE
@@ -19,6 +19,9 @@ export(ogs6_get_medium_property)
 export(ogs6_prj_top_level_classes)
 export(ogs6_read_output_files)
 export(ogs6_run_simulation)
+export(prj_basis_vector_0)
+export(prj_basis_vector_1)
+export(prj_basis_vector_2)
 export(prj_borehole_heat_exchanger)
 export(prj_boundary_condition)
 export(prj_capillary_pressure)
@@ -38,6 +41,7 @@ export(prj_fracture_model)
 export(prj_fracture_properties)
 export(prj_geometrical_set)
 export(prj_global_process_coupling)
+export(prj_ice_constitutive_relation)
 export(prj_insitu)
 export(prj_jacobian_assembler)
 export(prj_kinetic_reactant)
@@ -48,6 +52,7 @@ export(prj_medium)
 export(prj_micro_porosity)
 export(prj_nonlinear_solver)
 export(prj_output)
+export(prj_output_mesh)
 export(prj_parameter)
 export(prj_ph_property)
 export(prj_phase)
@@ -59,6 +64,7 @@ export(prj_pr_property)
 export(prj_process)
 export(prj_process_variable)
 export(prj_python_script)
+export(prj_raster)
 export(prj_rate)
 export(prj_relative_permeability)
 export(prj_search_length_algorithm)
