### 2024-11-17 cb0f07a
[doc] define versions of numpy, vtk and ogs in README
