### 2024-11-19 845bae1
[doc] protocol to update to ogs v6.5.2
