+++ update_protocols/protocol_update_v6.5.2_keywords.R
@@ -1,5 +1,5 @@
-# !!!!! READ the developer guide before updateing r2ogs6 !!!!!!
-# https://gitlab.opengeosys.org/ogs/tools/r2ogs6/-/blob/master/vignettes/dev_workflow_vignette.Rmd
+# !!!!! READ the developer guide before updating r2ogs6 !!!!!!
+# ./vignettes/dev_workflow_vignette.Rmd
 
 library(dplyr)
 devtools::load_all(".") # load current r2ogs6 version
@@ -48,16 +48,15 @@ generate_helper(raster, print_result = TRUE)
 
 # new class prj_raster added
 
-# execute data_raw/xpaths_for_classes.R
-xpaths_for_classes[["prj_raster"]]
+# execute code in data_raw/xpaths_for_classes.R
+xpaths_for_classes[["prj_raster"]] should show /OpenGeoSys/rasters/raster
 
 # add test with bm Parabolic/LiquidFlow/RasterParameter/GroundwaterModelWithRasterBC.prj
-
-
-
-
-
-
+# this bm does not work as the references raster file is not shipped with
+test <- OGS6$new("test", sim_path = paste0(temppath, "/test"))
+bm <- paste0("Parabolic/LiquidFlow/RasterParameter/GroundwaterModelWithRasterBC.prj")
+read_in_prj(test, paste0(temppath,"/Data_652/", bm), read_includes=TRUE)
+# read of tag works!
 
 
 
@@ -176,15 +175,17 @@ analyse_xml(path = paste0(temppath,"/Data_652"),
 
 
 # test with bm Parabolic/ComponentTransport/ReactiveTransport/DecayChain/1d_decay_chain_OS.prj
-# rm(test)
-# test <- OGS6$new("test", sim_path = paste0(temppath, "/test"))
-# bm <- paste0("Parabolic/LiquidFlow/SimpleSynthetics/PrimaryVariableConstraintDirichletBC/cuboid_1x1x1_hex_1000_Dirichlet_Dirichlet_1.prj")
-# read_in_prj(test, paste0(temppath,"/Data_644/", bm))
-# test$processes$LiquidFlow$linear
+rm(test)
+test <- OGS6$new("test", sim_path = paste0(temppath, "/test"))
+bm <- paste0("Parabolic/ComponentTransport/ReactiveTransport/DecayChain/1d_decay_chain_OS.prj")
+read_in_prj(test, paste0(temppath,"/Data_652/", bm))
+test$time_loop$output$output_extrapolation_residuals
 # # read of tag works!
-# dir_make_overwrite(test$sim_path)
-# run_benchmark(prj_path = paste0(temppath,"/Data_644/", bm),
-#               sim_path = test$sim_path)
+
+dir_make_overwrite(test$sim_path)
+run_benchmark(prj_path = paste0(temppath,"/Data_644/", bm),
+              sim_path = test$sim_path)
+# bm works!
 
 
 
@@ -205,12 +206,24 @@ analyse_xml(path = paste0(temppath,"/Data_652"),
             print_findings = TRUE)
 
 # class added to prj_process
+# added "steepness" and "characteristic_temperature" to class prj_pr_property
 
 # test with ThermoHydroMechanics/1D_freezing_column_Stefan/Stefan_problem.prj
+# rm(test)
+test <- OGS6$new("test", sim_path = paste0(temppath, "/test"))
+bm <- "ThermoHydroMechanics/1D_freezing_column_Stefan/Stefan_problem.prj"
[…]
