### 2024-11-20 c9f4540
[ci] update rocker, ogs and py module version
