### 2024-11-20 877c874
[base] set version for numpy and vtk for install ogs
