+++ R/config_ogs.R
@@ -5,16 +5,20 @@
 #' packages `vtk` and `numpy` into a specified Python virtual environment.
 #' Note, this does not work with `conda` environments.
 #'
-#' @param ogs_version The version of OGS to install (default: 6.4.4).
+#' @param ogs_version The version of OGS to install (default: 6.5.2).
 #' @param envname The name of the Python virtual environment (default: r2ogs6).
+#' @param numpy_version The version of `numpy` to install (default: 1.26.4).
+#' @param vtk_version The version of `vtk` to install (default: 9.3.1).
 #' @param ... Rest of the arguments passed to "py_install()"
 #' @return None.
-#'
+#',
 #' @export
 install_ogs <-
     function(...,
-           ogs_version = "6.4.4",
-           envname = "r2ogs6"
+            ogs_version = "6.5.2",
+            envname = "r2ogs6",
+            numpy_version = "1.26.4",
+            vtk_version = "9.3.1"
     ) {
 
     if (!(reticulate::virtualenv_exists(envname))){
@@ -28,7 +32,11 @@ install_ogs <-
         )
     }
 
-    pkgs <- c(paste0("ogs==", ogs_version), "numpy", "vtk")
+    pkgs <- c(
+        paste0("ogs==", ogs_version),
+        paste0("numpy==", numpy_version),
+        paste0("vtk==", vtk_version)
+    )
 
     reticulate::py_install(
       packages = pkgs,
