### 2024-11-20 545ad8f
[doc] update install_ogs fun doc
