+++ man/install_ogs.Rd
@@ -6,17 +6,28 @@
 packages `vtk` and `numpy` into a specified Python virtual environment.
 Note, this does not work with `conda` environments.}
 \usage{
-install_ogs(..., ogs_version = "6.4.4", envname = "r2ogs6")
+install_ogs(
+  ...,
+  ogs_version = "6.5.2",
+  envname = "r2ogs6",
+  numpy_version = "1.26.4",
+  vtk_version = "9.3.1"
+)
 }
 \arguments{
 \item{...}{Rest of the arguments passed to "py_install()"}
 
-\item{ogs_version}{The version of OGS to install (default: 6.4.4).}
+\item{ogs_version}{The version of OGS to install (default: 6.5.2).}
 
 \item{envname}{The name of the Python virtual environment (default: r2ogs6).}
+
+\item{numpy_version}{The version of `numpy` to install (default: 1.26.4).}
+
+\item{vtk_version}{The version of `vtk` to install (default: 9.3.1).}
 }
 \value{
 None.
+,
 }
 \description{
 Install specific version of OpenGeoSys 6 (OGS) along with necessary Python
