### 2024-11-20 d770b3c
[test]  specific numpy and vtk version for install_ogs
