### 2024-11-20 dad5429
[ci] install covr for stage test
