### 2024-11-24 51cd4bc
[doc] update version to 0.4.652
