### 2024-11-27 d7a4040
[bugfix] hdf5 bm path in user vignette
