### 2024-12-10 #5360 @xuanxu
This command should do it
