### 2024-12-10 #5360 @kthyng
@xuanxu Thank you! I didn't realize there is a bot command for that now.
