### 2024-12-10 #5360 @kthyng
@joboog Please address the doi errors ⬆️

Also please check the capitalization in your references. You can preserve capitalization by placing {} around characters/words in your .bib file. For example "python", "r", "Latin hypercube samples" but please check for more.
