### 2024-12-17 b800698
[JOSS] fix DOIs
