+++ JOSS/paper.bib
@@ -1,7 +1,7 @@
 @article{Field2012,
 author={Field, C. and Barros, V. and Stocker, T. and Dahe, Q.},
 title={Managing the Risks of Extreme Events and Disasters to Advance Climate Change Adaptation: Special Report of the Intergovernmental Panel on Climate Change},
-DOI={10.1017/CBO9781139177245},
+doi={10.1017/CBO9781139177245},
 publisher={Cambridge University Press},
 year={2012}}
 
@@ -12,7 +12,8 @@ year={2012}}
   volume={553},
   pages={105--118},
   year={2017},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/j.jhydrol.2017.07.048}
 }
 
 @article{Wu2020,
@@ -23,7 +24,8 @@ year={2012}}
   number={1},
   pages={3710},
   year={2020},
-  publisher={Nature Publishing Group UK London}
+  publisher={Nature Publishing Group UK London},
+  doi={10.1038/s41467-020-17581-y}
 }
 
 @article{Li2023,
@@ -32,7 +34,8 @@ year={2012}}
   journal={Nature Reviews Earth \& Environment},
   pages={1--14},
   year={2023},
-  publisher={Nature Publishing Group UK London}
+  publisher={Nature Publishing Group UK London},
+  doi={10.1038/s43017-023-00409-w}
 }
 %------------------ OpenGeoSys
 
@@ -47,7 +50,8 @@ number = {3},
 pages = {445-478},
 organization = {Lawrence Berkeley National Lab. (LBNL), Berkeley, CA (United States)},
 publisher = {Springer International Publishing},
-address = {Cham}
+address = {Cham},
+doi={10.1007/s10596-014-9443-x}
 }
 
 @article{Bilke2019,
@@ -94,7 +98,7 @@ volume = {178},
 pages = {46-61},
 year = {2020},
 issn = {0378-4754},
-doi = {https://doi.org/10.1016/j.matcom.2020.05.024},
+doi = {10.1016/j.matcom.2020.05.024},
 url = {https://www.sciencedirect.com/science/article/pii/S037847542030183X},
 author = {Norbert Grunwald and Jobst Maßmann and Olaf Kolditz and Thomas Nagel},
 keywords = {Phase equilibrium, CO-HO-NaCl-system, Saline aquifer, CO storage, OpenGeoSys},
@@ -108,7 +112,8 @@ abstract = {This article features a simple method for describing the phase equil
   volume={11},
   pages={1061420},
   year={2023},
-  publisher={Frontiers Media SA}
+  publisher={Frontiers Media SA},
+  doi={10.3389/feart.2023.1061420}
 }
 
 @article{Li2014,
@@ -118,7 +123,8 @@ abstract = {This article features a simple method for describing the phase equil
   volume={9},
   pages={67--79},
   year={2014},
-  publisher={Springer}
+  publisher={Springer},
+  doi={10.1007/s11440-013-0234-7}
 }
 
 @article{Jing2018,
@@ -129,7 +135,8 @@ abstract = {This article features a simple method for describing the phase equil
   number={5},
   pages={1989--2007},
   year={2018},
-  publisher={Copernicus GmbH}
+  publisher={Copernicus GmbH},
+  doi={10.5194/gmd-11-1989-2018}
 }
 
 @article{Pitz2023,
@@ -140,7 +147,8 @@ abstract = {This article features a simple method for describing the phase equil
   number={13},
   pages={1--24},
   year={2023},
-  publisher={Springer}
+  publisher={Springer},
+  doi={10.1007/s12665-023-10971-7}
 }
 
 @article{Asher2015,
[…]
