### 2024-12-17 #5360 @joboog
@kthyng All missing and invalid DOIs have been fixed.
