### 2024-12-17 6488879
[JOSS] fix capitalization in bibliography
