+++ JOSS/paper.bib
@@ -6,7 +6,7 @@ publisher={Cambridge University Press},
 year={2012}}
 
 @article{Felfelani2017,
-  title={Natural and human-induced terrestrial water storage change: A global analysis using hydrological models and GRACE},
+  title={Natural and human-induced terrestrial water storage change: A global analysis using hydrological models and {GRACE}},
   author={Felfelani, Farshid and Wada, Yoshihide and Longuevergne, Laurent and Pokhrel, Yadu N},
   journal={Journal of Hydrology},
   volume={553},
@@ -87,12 +87,12 @@ doi={10.1007/s10596-014-9443-x}
   number = {67},
   pages = {3673},
   author = {J{\"o}rg Buchwald and Olaf Kolditz and Thomas Nagel},
-  title = {ogs6py and VTUinterface: streamlining OpenGeoSys workflows in Python},
+  title = {{ogs6py and VTUinterface}: streamlining {OpenGeoSys} workflows in {Python}},
   journal = {Journal of Open Source Software}
 }
 
 @article{Grunwald2020,
-title = {Non-iterative phase-equilibrium model of the H2O-CO2-NaCl-system for large-scale numerical simulations},
+title = {Non-iterative phase-equilibrium model of the {H2O-CO2-NaCl-system} for large-scale numerical simulations},
 journal = {Mathematics and Computers in Simulation},
 volume = {178},
 pages = {46-61},
@@ -106,7 +106,7 @@ abstract = {This article features a simple method for describing the phase equil
 }
 
 @article{Pujades2023,
-  title={Towards the construction of representative regional hydro (geo) logical numerical models: Modelling the upper Danube basin as a starting point},
+  title={Towards the construction of representative regional hydro (geo) logical numerical models: {Modelling} the upper {Danube} basin as a starting point},
   author={Pujades, Estanislao and Kumar, Rohini and Houben, Timo and Jing, Miao and Rakovec, Oldrich and Kalbacher, Thomas and Attinger, Sabine},
   journal={Frontiers in Earth Science},
   volume={11},
@@ -117,7 +117,7 @@ abstract = {This article features a simple method for describing the phase equil
 }
 
 @article{Li2014,
-  title={OpenGeoSys-ChemApp: a coupled simulator for reactive transport in multiphase systems and application to CO 2 storage formation in Northern Germany},
+  title={{OpenGeoSys-ChemApp}: a coupled simulator for reactive transport in multiphase systems and application to CO 2 storage formation in Northern {Germany}},
   author={Li, Dedong and Bauer, Sebastian and Benisch, Katharina and Graupner, Bastian and Beyer, Christof},
   journal={Acta Geotechnica},
   volume={9},
@@ -128,7 +128,7 @@ abstract = {This article features a simple method for describing the phase equil
 }
 
 @article{Jing2018,
-  title={Improved regional-scale groundwater representation by the coupling of the mesoscale Hydrologic Model (mHM v5. 7) to the groundwater model OpenGeoSys (OGS)},
+  title={Improved regional-scale groundwater representation by the coupling of the {mesoscale Hydrologic Model (mHM v5. 7)} to the groundwater model {OpenGeoSys (OGS)}},
   author={Jing, Miao and He{\ss}e, Falk and Kumar, Rohini and Wang, Wenqing and Fischer, Thomas and Walther, Marc and Zink, Matthias and Zech, Alraune and Samaniego, Luis and Kolditz, Olaf and others},
   journal={Geoscientific Model Development},
   volume={11},
[…]
