### 2024-12-18 #5360 @kthyng
Congratulations on your new publication @joboog, and sorry for the very extended time frame for this submission! Many thanks to reviewers @waternumbers and @ldecicco-USGS for your time, hard work, and expertise!! JOSS wouldn't be able to function nor succeed without your efforts.
